backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
youtube-transcript-api==0.6.2
yt-dlp==2024.5.27
pydantic==2.7.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0